% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicMultiJGLplot.R
\name{dynamicMultiJGLplot}
\alias{dynamicMultiJGLplot}
\title{Title dynamic plot creation function}
\usage{
dynamicMultiJGLplot(
  networks,
  display.networks = c("LINEAR"),
  display.unique = "FALSE",
  main = NULL,
  xlab = NULL,
  output.mode = "htmlWidget",
  ...
)
}
\arguments{
\item{networks}{Estimated dynamic network structure}

\item{display.networks}{Whether LINEAR, NONLINEAR or COMBINED networks are considered}

\item{display.unique}{Whether all structures or only group specific unique structures are plotted}

\item{main}{Title for the animation}

\item{xlab}{xlab for the animation}

\item{output.mode}{output mode}

\item{...}{additional arguments for the render.d3movie function used to create animations}
}
\value{
Dynamic output, e.g., as htmlWidget
}
\description{
Title dynamic plot creation function
}
\examples{
print("dynamicMultiJGLplot(dynamicMultiJGL(x))")
}
