% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiJGLplot.R
\name{multiJGLplot}
\alias{multiJGLplot}
\title{Visualisation function for static multiclass network models}
\usage{
multiJGLplot(
  x,
  obs.class.names = NULL,
  node.names = NULL,
  network.type = c("linear", "nonlinear", "linear and nonlinear"),
  structures.to.plot = "obs.class",
  obs.class.legend = TRUE,
  obs_class.index = NULL,
  graphlayout = NULL,
  lcex = 0.5,
  ...
)
}
\arguments{
\item{x}{Estimated network structures from the static_LNLJGL function.}

\item{obs.class.names}{Observational class names.}

\item{node.names}{Network node names.}

\item{network.type}{Should linear, nonlinear or combined networks be plotted.}

\item{structures.to.plot}{Should obs.class specific structures to be specified.}

\item{obs.class.legend}{Optional: Obs class specific legends.}

\item{obs_class.index}{Optional: Used to specify the subnetworks.}

\item{graphlayout}{igraph argument.}

\item{lcex}{the cex argument for legends.}

\item{...}{Additional igraph arguments.}
}
\value{
Multiclass network plot.
}
\description{
Visualisation function for static multiclass network models
}
\examples{
print("multiJGLplot(multiJGLoutput$linear.strs)")
}
