% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xenaprep.R
\name{Xenaprep}
\alias{Xenaprep}
\title{Omics data downloading and preprocessing function}
\usage{
Xenaprep(
  cohort = "TCGA",
  cancertype = "LAML",
  type = c("clinical", "molecular", "mRNA"),
  merge = TRUE
)
}
\arguments{
\item{cohort}{Cohort abbreviation from Xenahub.}

\item{cancertype}{Specific tumor TCGA abbreviation (e.g., "LAML", "BRCA").}

\item{type}{type to be downloaded and merged including "clinical" "molecular" "mRNA".}

\item{merge}{If multiple values are given for the type argument, should resulting datasets be merged by sampleID.}
}
\description{
Omics data downloading and preprocessing function
}
\examples{
print("")
}
