% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiJGLplot.R
\name{multiJGLplot}
\alias{multiJGLplot}
\title{Visualisation function for static multiclass network models}
\usage{
multiJGLplot(
  x,
  obs.class.names = NULL,
  node.names = NULL,
  obs.class.legend = TRUE,
  graphlayout = NULL,
  lcex = 0.5,
  ...
)
}
\arguments{
\item{x}{An individual element (e.g. $linear.strs) from the output of multiJGL function.}

\item{obs.class.names}{Observational class names.}

\item{node.names}{Network node names.}

\item{obs.class.legend}{Optional: Obs class specific legends.}

\item{graphlayout}{igraph argument.}

\item{lcex}{the cex argument for legends.}

\item{...}{Additional igraph arguments.}
}
\value{
Multiclass network plot.
}
\description{
Visualisation function for static multiclass network models
}
\examples{
print("multiJGLplot(multiJGLoutput$linear.strs)")
}
